pyproject.toml
[project]
name = "ForeAIScripts"
version = "0.1.0"
description = "ForeAI automation and test-data scripts"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["."]
exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.ruff.format]
quote-style = "double"
